requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
celery>=5.3.0
redis>=4.5.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
httpx>=0.25.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
dj-database-url>=2.1.0
responses>=0.25.0
drf-spectacular>=0.26.5
ffmpeg-python>=0.2.0
google-cloud-storage>=2.18.2
